# src/BSDCPolls.BFF/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BSDCPolls BFF",
    "version": "v1"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "operationId": "Auth_Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "Auth_Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/polls/{pollUid}/invitations": {
      "post": {
        "tags": ["Invitations"],
        "operationId": "Invitations_CreatePollInvitation",
        "parameters": [
          { "name": "pollUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateInvitationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvitationResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
    "/api/surveys/{surveyUid}/invitations": {
      "post": {
        "tags": ["Invitations"],
        "operationId": "Invitations_CreateSurveyInvitation",
        "parameters": [
          { "name": "surveyUid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateInvitationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvitationResponse" }
              }
            }
          },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        },
        "security": [{ "Bearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "password": { "type": "string", "minLength": 12 }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "required": ["username", "userUid"],
        "properties": {
          "username": { "type": "string" },
          "userUid": { "type": "string", "format": "uuid" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "maxLength": 60 },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["accessToken", "expiresAt", "userUid", "username"],
        "properties": {
          "accessToken": { "type": "string" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "userUid": { "type": "string", "format": "uuid" },
          "username": { "type": "string" }
        }
      },
...
      "CreateInvitationRequest": {
        "type": "object",
        "required": ["targetUsername"],
        "properties": {
          "targetUsername": { "type": "string", "maxLength": 60 }
        }
      },
      "InvitationResponse": {
        "type": "object",
        "required": ["invitationUid", "targetUsername", "targetUserUid", "createdOn"],
        "properties": {
          "invitationUid": { "type": "string", "format": "uuid" },
          "targetUsername": { "type": "string" },
          "targetUserUid": { "type": "string", "format": "uuid" },
          "createdOn": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}